requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
opencv-python==4.10.0.0
opencv-contrib-python==4.10.0.0
numpy==1.23.5
paho-mqtt==1.6.1
